move _MutableArrayData
