+++ arrow-data/src/transform/mod.rs
@@ -24,6 +24,7 @@ use half::f16;
 use num::Integer;
 use std::mem;
 pub use traits::SpecializedMutableArrayData;
+use crate::transform::utils::{_MutableArrayData, build_extend_null_bits};
 
 mod boolean;
 mod fixed_binary;
@@ -44,60 +45,6 @@ type Extend<'a> = Box<dyn Fn(&mut _MutableArrayData, usize, usize, usize) + 'a>;
 
 type ExtendNulls = Box<dyn Fn(&mut _MutableArrayData, usize)>;
 
-/// A mutable [ArrayData] that knows how to freeze itself into an [ArrayData].
-/// This is just a data container.
-#[derive(Debug)]
-struct _MutableArrayData<'a> {
-    pub data_type: DataType,
-    pub null_count: usize,
-
-    pub len: usize,
-    pub null_buffer: Option<MutableBuffer>,
-
-    // arrow specification only allows up to 3 buffers (2 ignoring the nulls above).
-    // Thus, we place them in the stack to avoid bound checks and greater data locality.
-    pub buffer1: MutableBuffer,
-    pub buffer2: MutableBuffer,
-    pub child_data: Vec<MutableArrayData<'a>>,
-}
-
-impl<'a> _MutableArrayData<'a> {
-    fn null_buffer(&mut self) -> &mut MutableBuffer {
-        self.null_buffer
-            .as_mut()
-            .expect("MutableArrayData not nullable")
-    }
-}
-
-fn build_extend_null_bits(array: &ArrayData, use_nulls: bool) -> ExtendNullBits {
-    if let Some(nulls) = array.nulls() {
-        let bytes = nulls.validity();
-        Box::new(move |mutable, start, len| {
-            let mutable_len = mutable.len;
-            let out = mutable.null_buffer();
-            utils::resize_for_bits(out, mutable_len + len);
-            mutable.null_count += set_bits(
-                out.as_slice_mut(),
-                bytes,
-                mutable_len,
-                nulls.offset() + start,
-                len,
-            );
-        })
-    } else if use_nulls {
-        Box::new(|mutable, _, len| {
-            let mutable_len = mutable.len;
-            let out = mutable.null_buffer();
-            utils::resize_for_bits(out, mutable_len + len);
-            let write_data = out.as_slice_mut();
-            (0..len).for_each(|i| {
-                bit_util::set_bit(write_data, mutable_len + i);
-            });
-        })
-    } else {
-        Box::new(|_, _, _| {})
-    }
-}
 
 /// Efficiently create an [ArrayData] from one or more existing [ArrayData]s by
 /// copying chunks.
